// rent-prompts-ai-adding-more-clients/src/payload-types.ts
export interface Model {
  id: string;
  name: string;
  description: string;
  type: 'text' | 'image' | 'audio' | 'video';
  imageinput: boolean;
  provider: 'groq' | 'openai' | 'replicate' | 'falai';
  settings: {
    name: string;
    type: 'integer' | 'string' | 'boolean' | 'select';
    description: string;
    options?: string[] | null;
    allowMultiple: boolean;
    id?: string | null;
  }[];
  enablePrompt: boolean;
  systemprompt: boolean;
  negativeprompt: boolean;
  prodkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apikey: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
    falai?: {
      modelname: string;
      apikey: string;
    };
  };
  testkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apikey: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
    falai?: {
      modelname: string;
      apikey: string;
    };
  };
  cost: number;
  commissionapplicable: boolean;
  commision?: number | null;
  tokens: number;
  examples: {
    example: string;
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
}